# [drci] Improve commit_jobs_red query perf, dedup base sha list