--- a/torchci/clickhouse_queries/commit_failed_jobs/params.json
+++ b/torchci/clickhouse_queries/commit_failed_jobs/params.json
@@ -2,5 +2,34 @@
   "params": {
     "shas": "String"
   },
-  "tests": []
+  "tests": [
+    {
+      "shas": [
+        "53f73cdeb6fe53155f91064ca15b722e80dec2f3",
+        "7f88bf96f9ff37f3873cc5cb8c962ebcbf509c65",
+        "0c139fa58e7586a77233460aeda3427652c9fc95",
+        "aade4fbd55a07aaa23dbdfe055d70cd503fd0059",
+        "a89bdc0565b9026bd17afab2424ac857c39cead8",
+        "2a7e997b3f1805b077810d7fef87cabc4411eeea",
+        "87bfd66c3c7061db6d36d8daa62f08f507f90e39",
+        "23183fef7ec07ea00015655a39472c42a2dcd78e",
+        "ce54c430c0e9d5e6e9ee0b1d85bddd04fbcbca4e",
+        "5beb5b7e47750d63f42684b9411ff929b07894c9",
+        "7336b76bccc5673221d4a8296b93d2ed7229b48c",
+        "de3aca331104776bdfac3f053511b7af442b4038",
+        "9bae904cb47b2d896f4653f751f0526379823606",
+        "5a7588f1832a840285ed29b039f01b9031570e5c",
+        "039ebdc19287bba56fdb6d6bb9e693b3c88de927",
+        "819b23e0b43af2f3417d48fc84643fad2e3133f1",
+        "7379c663441dc05b5c263075e9f19ead2628740f",
+        "7ca296f564e06afda55843ab7583d15cbfea443a",
+        "23855391f1a17f7145885b5ef977547a70819505",
+        "248487f455e943cbba368404119ca9bcb14c0499",
+        "924a247fbbe32063966533535b1cc0cf474ef1ed",
+        "4a9466c96a97657e83e1ba418b942f73a4213afc",
+        "8d08b4901586f230353a558ee00c16ad57f95178",
+        "b248edd7ccae15cf3a2dd86442149e4bc029846a"
+      ]
+    }
+  ]
 }