--- a/torchci/clickhouse_queries/commit_failed_jobs/query.sql
+++ b/torchci/clickhouse_queries/commit_failed_jobs/query.sql
@@ -1,25 +1,26 @@
 -- This query is used by Dr.CI to get all the failed jobs from the base commit. They can then be
 -- used to decide if a failure is due to broken trunk
-with relevant_pushes as (
+with runs as (
   select
-    p.head_commit.timestamp as timestamp,
-    p.head_commit.'id' as after,
-    p.head_commit.'author'.'email' as authorEmail
-  from default.push p
+    w.id as id,
+    w.head_commit.'author'.'email' as authorEmail,
+    w.head_commit.'id' as head_sha,
+    w.head_commit.timestamp as head_sha_timestamp
+  from default.workflow_run w
   where
-    p.head_commit.'id' in {shas: Array(String)}
+    w.id in (select id from materialized_views.workflow_run_by_head_sha where head_sha in {shas: Array(String)})
 )
 SELECT
   j.id as id,
   j.name AS jobName,
   CONCAT(j.workflow_name, ' / ', j.name) AS name,
   j.runner_name AS runnerName,
-  p.authorEmail as authorEmail,
+  w.authorEmail as authorEmail,
   j.conclusion as conclusion,
   j.completed_at as completed_at,
   j.html_url as html_url,
   j.head_sha as head_sha,
-  p.timestamp AS head_sha_timestamp,
+  w.head_sha_timestamp AS head_sha_timestamp,
   j.head_branch as head_branch,
   j.torchci_classification.'captures' AS failure_captures,
   IF(j.torchci_classification.'line' = '', [], [j.torchci_classification.'line']) AS failure_lines,
@@ -27,7 +28,7 @@ SELECT
   j.created_at AS time
 FROM
   default.workflow_job j final
-  JOIN relevant_pushes p ON p.after = j.head_sha
+  JOIN runs w on w.id = j.run_id
 WHERE
   j.id in (select id from materialized_views.workflow_job_by_head_sha where head_sha in {shas: Array(String)})
   AND j.conclusion IN ('failure', 'cancelled')