--- a/torchci/pages/api/drci/drci.ts
+++ b/torchci/pages/api/drci/drci.ts
@@ -75,7 +75,7 @@ export default async function handler(
 ) {
   const authorization = req.headers.authorization;
 
-  if (true || authorization == process.env.DRCI_BOT_KEY) {
+  if (authorization == process.env.DRCI_BOT_KEY) {
     // Dr. CI bot key is used to update the comment, probably called from the
     // update Dr. CI workflow
   } else if (authorization) {
@@ -141,7 +141,6 @@ export async function updateDrciComments(
   const unstableIssues: IssueData[] = await fetchIssuesByLabel("unstable");
   const disabledTestIssues: IssueData[] = await fetchIssuesByLabel("skipped");
   const baseCommitJobs = await getBaseCommitJobs(workflowsByPR);
-  return {};
   const existingDrCiComments = await getExistingDrCiComments(
     `${OWNER}/${repo}`,
     workflowsByPR
@@ -440,7 +439,6 @@ export async function getBaseCommitJobs(
   const baseShas = _.uniq(
     Array.from(workflowsByPR.values()).map((v) => v.merge_base)
   );
-  console.log(baseShas)
 
   // fetch failing jobs on those shas
   const commitFailedJobsQueryResult = await fetchFailedJobsFromCommits(